// app/components/theme-provider/theme/types.ts
export type TColor =
  | "color.white"
  | "color.gray.1"
  | "color.gray.2"
  | "color.gray.3"
  | "color.gray.4"
  | "color.gray.5"
  | "color.gray.6"
  | "color.gray.7"
  | "color.gray.8"
  | "color.gray.9"
  | "color.gray.10"
  | "color.blue"
  | "color.blue.1"
  | "color.blue.2"
  | "color.blue.3"
  | "color.blue.4"
  | "color.blue.5"
  | "color.blue.6"
  | "color.blue.7"
  | "color.blue.8"
  | "color.blue.9"
  | "color.blue.10"
  | "color.green"
  | "color.green.1"
  | "color.green.2"
  | "color.green.3"
  | "color.green.4"
  | "color.green.5"
  | "color.green.6"
  | "color.green.7"
  | "color.green.8"
  | "color.green.9"
  | "color.green.10"
  | "color.cyan"
  | "color.cyan.1"
  | "color.cyan.2"
  | "color.cyan.3"
  | "color.cyan.4"
  | "color.cyan.5"
  | "color.cyan.6"
  | "color.cyan.7"
  | "color.cyan.8"
  | "color.cyan.9"
  | "color.cyan.10"
  | "color.yellow"
  | "color.yellow.1"
  | "color.yellow.2"
  | "color.yellow.3"
  | "color.yellow.4"
  | "color.yellow.5"
  | "color.yellow.6"
  | "color.yellow.7"
  | "color.yellow.8"
  | "color.yellow.9"
  | "color.yellow.10"
  | "color.orange"
  | "color.orange.1"
  | "color.orange.2"
  | "color.orange.3"
  | "color.orange.4"
  | "color.orange.5"
  | "color.orange.6"
  | "color.orange.7"
  | "color.orange.8"
  | "color.orange.9"
  | "color.orange.10"
  | "color.red"
  | "color.red.1"
  | "color.red.2"
  | "color.red.3"
  | "color.red.4"
  | "color.red.5"
  | "color.red.6"
  | "color.red.7"
  | "color.red.8"
  | "color.red.9"
  | "color.red.10"
  | "color.purple"
  | "color.purple.1"
  | "color.purple.2"
  | "color.purple.3"
  | "color.purple.4"
  | "color.purple.5"
  | "color.purple.6"
  | "color.purple.7"
  | "color.purple.8"
  | "color.purple.9"
  | "color.purple.10"
  | "color.pink"
  | "color.pink.1"
  | "color.pink.2"
  | "color.pink.3"
  | "color.pink.4"
  | "color.pink.5"
  | "color.pink.6"
  | "color.pink.7"
  | "color.pink.8"
  | "color.pink.9"
  | "color.pink.10"
  | "color.primary"
  | "color.primary.1"
  | "color.primary.2"
  | "color.primary.3"
  | "color.primary.4"
  | "color.primary.5"
  | "color.primary.6"
  | "color.primary.7"
  | "color.primary.8"
  | "color.primary.9"
  | "color.primary.10"
  | "color.success"
  | "color.success.1"
  | "color.success.2"
  | "color.success.3"
  | "color.success.4"
  | "color.success.5"
  | "color.success.6"
  | "color.success.7"
  | "color.success.8"
  | "color.success.9"
  | "color.success.10"
  | "color.info"
  | "color.info.1"
  | "color.info.2"
  | "color.info.3"
  | "color.info.4"
  | "color.info.5"
  | "color.info.6"
  | "color.info.7"
  | "color.info.8"
  | "color.info.9"
  | "color.info.10"
  | "color.warning"
  | "color.warning.1"
  | "color.warning.2"
  | "color.warning.3"
  | "color.warning.4"
  | "color.warning.5"
  | "color.warning.6"
  | "color.warning.7"
  | "color.warning.8"
  | "color.warning.9"
  | "color.warning.10"
  | "color.error"
  | "color.error.1"
  | "color.error.2"
  | "color.error.3"
  | "color.error.4"
  | "color.error.5"
  | "color.error.6"
  | "color.error.7"
  | "color.error.8"
  | "color.error.9"
  | "color.error.10";

export type TBackground =
  | "color.background.body"
  | "color.background.component";

export type TLink = "color.link" | "color.link.hover" | "color.link.active";

export type TSpacing =
  | "spacing"
  | "spacing.1"
  | "spacing.2"
  | "spacing.3"
  | "spacing.4"
  | "spacing.5"
  | "spacing.6"
  | "spacing.7"
  | "spacing.8"
  | "spacing.9"
  | "spacing.10";

export type TFont =
  | "font.family"
  | "font.size"
  | "font.weight"
  | "font.size.small"
  | "font.size.large"
  | "font.color"
  | "font.color.secondary"
  | "line.height";

export type TBorder =
  | "border.color"
  | "border.color.secondary"
  | "border.radius"
  | "border.radius.small"
  | "border.radius.large";

export type TVariables = Partial<
  Record<TColor | TBackground | TSpacing | TFont | TBorder | TLink, string>
>;